requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
sqlalchemy==2.0.30
aiomysql==0.2.0
PyMySQL==1.1.1
fastapi-users[sqlalchemy]==13.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
azure-storage-blob==12.19.1
pydantic-settings==2.2.1
alembic==1.13.1
python-dotenv==1.0.1
